# docs/v3/openapi3.json
{
	"components": {
		"schemas": {
...
			"datastore.CustomResponse": {
				"properties": {
					"body": {
						"type": "string"
					},
					"content_type": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"datastore.PaginationData": {
				"properties": {
					"has_next_page": {
						"type": "boolean"
					},
					"has_prev_page": {
						"type": "boolean"
					},
					"next_page_cursor": {
						"type": "string"
					},
					"per_page": {
						"type": "integer"
					},
					"prev_page_cursor": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"handlers.Stub": {
				"type": "object"
			},
...
			"models.CreateSource": {
				"properties": {
					"body_function": {
						"description": "Function is a javascript function used to mutate the payload\nimmediately after ingesting an event",
						"type": "string"
					},
					"custom_response": {
						"allOf": [
							{
								"$ref": "#/components/schemas/models.CustomResponse"
							}
						],
						"description": "Custom response is used to define a custom response for incoming\nwebhooks project sources only."
					},
					"header_function": {
						"description": "Function is a javascript function used to mutate the headers\nimmediately after ingesting an event",
						"type": "string"
					},
					"idempotency_keys": {
						"description": "IdempotencyKeys are used to specify parts of a webhook request to uniquely\nidentify the event in an incoming webhooks project.",
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"name": {
						"description": "Source name.",
						"type": "string"
					},
					"provider": {
						"allOf": [
							{
								"$ref": "#/components/schemas/datastore.SourceProvider"
							}
						],
						"description": "Use this to specify one of our predefined source types."
					},
					"pub_sub": {
						"allOf": [
							{
								"$ref": "#/components/schemas/models.PubSubConfig"
							}
						],
						"description": "PubSub are used to specify message broker sources for outgoing\nwebhooks projects."
					},
					"type": {
						"allOf": [
							{
								"$ref": "#/components/schemas/datastore.SourceType"
							}
						],
						"description": "Source Type."
					},
					"verifier": {
						"allOf": [
							{
								"$ref": "#/components/schemas/models.VerifierConfig"
							}
						],
						"description": "Verifiers are used to verify webhook events ingested in incoming\nwebhooks projects."
					}
				},
				"type": "object"
			},
...
			"models.FunctionRequest": {
				"properties": {
					"function": {
						"type": "string"
					},
					"payload": {
						"additionalProperties": {},
						"type": "object"
					},
					"type": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"models.FunctionResponse": {
				"properties": {
					"log": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"payload": {}
				},
				"type": "object"
			},
...
			"models.PagedResponse": {
				"properties": {
					"content": {},
					"pagination": {
						"$ref": "#/components/schemas/datastore.PaginationData"
					}
				},
				"type": "object"
			},
...
			"models.SourceResponse": {
				"properties": {
					"body_function": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"custom_response": {
						"$ref": "#/components/schemas/datastore.CustomResponse"
					},
					"deleted_at": {
						"type": "string"
					},
					"forward_headers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"header_function": {
						"type": "string"
					},
					"idempotency_keys": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"is_disabled": {
						"type": "boolean"
					},
					"mask_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"project_id": {
						"type": "string"
					},
					"provider": {
						"$ref": "#/components/schemas/datastore.SourceProvider"
					},
					"provider_config": {
						"$ref": "#/components/schemas/datastore.ProviderConfig"
					},
					"pub_sub": {
						"$ref": "#/components/schemas/datastore.PubSubConfig"
					},
					"type": {
						"$ref": "#/components/schemas/datastore.SourceType"
					},
					"uid": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"verifier": {
						"$ref": "#/components/schemas/datastore.VerifierConfig"
					}
				},
				"type": "object"
			},
...
			"util.ServerResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					}
				},
				"type": "object"
			}
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"in": "header",
				"name": "Authorization",
				"type": "apiKey"
			}
		}
	},
	"info": {
		"contact": {
			"email": "support@getconvoy.io",
			"name": "Convoy Support",
			"url": "https://getconvoy.io/docs"
		},
		"description": "Convoy is a fast and secure webhooks proxy. This document contains datastore.s API specification.",
		"license": {
			"name": "Mozilla Public License 2.0",
			"url": "https://www.mozilla.org/en-US/MPL/2.0/"
		},
		"termsOfService": "https://getconvoy.io/terms",
		"title": "Convoy API Reference",
		"version": "24.1.4"
	},
	"openapi": "3.0.3",
	"paths": {
...
		"/v1/projects/{projectID}/sources": {
			"get": {
				"description": "This endpoint fetches multiple sources",
				"operationId": "LoadSourcesPaged",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "direction",
						"schema": {
							"enum": [
								"next",
								"prev"
							],
							"type": "string"
						},
						"x-enum-varnames": [
							"Next",
							"Prev"
						]
					},
					{
						"description": "A pagination cursor to fetch the next page of a list",
						"in": "query",
						"name": "next_page_cursor",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The number of items to return per page",
						"in": "query",
						"name": "perPage",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "A pagination cursor to fetch the previous page of a list",
						"in": "query",
						"name": "prev_page_cursor",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The custom source provider e.g. twitter, shopify",
						"in": "query",
						"name": "provider",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Sort order, values are `ASC` or `DESC`, defaults to `DESC`",
						"in": "query",
						"name": "sort",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The source type e.g. http, pub_sub",
						"in": "query",
						"name": "type",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"allOf": [
														{
															"$ref": "#/components/schemas/models.PagedResponse"
														},
														{
															"properties": {
																"content": {
																	"items": {
																		"$ref": "#/components/schemas/models.SourceResponse"
																	},
																	"type": "array"
																}
															},
															"type": "object"
														}
													]
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "List all sources",
				"tags": [
					"Sources"
				]
			},
			"post": {
				"description": "This endpoint creates a source",
				"operationId": "CreateSource",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/models.CreateSource"
							}
						}
					},
					"description": "Source Details",
					"required": true,
					"x-originalParamName": "source"
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/models.SourceResponse"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Created"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "Create a source",
				"tags": [
					"Sources"
				]
			}
		},
		"/v1/projects/{projectID}/sources/{sourceID}": {
			"delete": {
				"description": "This endpoint deletes a source",
				"operationId": "DeleteSource",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "source id",
						"in": "path",
						"name": "sourceID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "Delete a source",
				"tags": [
					"Sources"
				]
			},
			"get": {
				"description": "This endpoint retrieves a source by its id",
				"operationId": "GetSource",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Source ID",
						"in": "path",
						"name": "sourceID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/models.SourceResponse"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "Retrieve a source",
				"tags": [
					"Sources"
				]
			},
			"put": {
				"description": "This endpoint updates a source",
				"operationId": "UpdateSource",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "source id",
						"in": "path",
						"name": "sourceID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/models.UpdateSource"
							}
						}
					},
					"description": "Source Details",
					"required": true,
					"x-originalParamName": "source"
				},
				"responses": {
					"202": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/models.SourceResponse"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Accepted"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "Update a source",
				"tags": [
					"Sources"
				]
			}
		},
...
		"/v1/projects/{projectID}/subscriptions/test_function": {
			"post": {
				"description": "This endpoint test runs a transform function against a payload.",
				"operationId": "TestSubscriptionFunction",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/models.FunctionRequest"
							}
						}
					},
					"description": "Function Details",
					"required": true,
					"x-originalParamName": "filter"
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/models.FunctionResponse"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "Test a subscription function",
				"tags": [
					"Subscriptions"
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://dashboard.getconvoy.io/api"
		}
	],
...
}